requirements.txt
djangorestframework==3.15.2
django-filter==24.3
psycopg>=3.1.0
psycopg2-binary==2.9.10; python_version < "3.13"
pypdf==5.1.0
python-dotenv==1.0.1
gunicorn==23.0.0
